Add 'New window' functionality to SimpleTimer.
This enables SimpleTimer to launch multiple windows within the same
instance of the application.

This is important in Mac OS X because an application can normally only
be launched in one instance at a time per user (you can force a launch
of a second instance through the command line, but that's not supported
and can cause other problems).